==> processing/00-get_data.Rmd <==
---
title: "00-get_data"
author: "Kevin Carrasco"
date: "2024-01-16"
output: html_document
---

# packages

```{r packages}
devtools::install_github("rubenarslan/formr")
pacman::p_load(formr, dplyr, haven, writexl)
```


# Descargar base desde formR

```{r}
formr_connect("kevin.carrasco@ug.uchile.cl", "k959371343") # credenciales FormR
data = formr_results("Encuesta") # Descargar encuesta
writexl::write_xlsx(data, path = "input/data/original/encuesta-hvs-completa.xlsx") # Guardar en xlsx
save(data, file="input/data/original/encuesta-hvs-completa.RData") # Guardar en RData
```

# Filtrar solo respuestas con información

```{r}
load("input/data/original/encuesta-hvs-completa.RData")
hvs <- data %>% filter(!is.na(carac_d1))
`%!in%` <- Negate(`%in%`)
hvs <- hvs %>% filter(carac_d1 %!in% c('a', 'Agustiniano', 'U mayor')) # eliminar dos casos de etapa prueba
hvs <- hvs %>% dplyr::select(-session, -created, -modified, -ended, -expired, -start_01, -hva, -ha, -ge)
#writexl::write_xlsx(hvs, path = "input/data/proc/encuesta-hvs-limpia.xlsx") # Guardar en xlsx y seleccionamos 'a mano'
#hvs<-readxl::read_xlsx("input/data/proc/encuesta-hvs-limpia.xlsx")
hvs$RBD <- sub("-.*", "", hvs$carac_d2)

completas <- hvs %>% filter(!is.na(ge_4))
completas <- completas[!duplicated(completas$RBD),]

hvs <- hvs[!duplicated(hvs$RBD),]
save(hvs, file="input/data/proc/encuesta_hvs.RData")
writexl::write_xlsx(hvs, path = "input/data/proc/encuesta_hvs.xlsx")
```
