==> processing/01-preparacion.Rmd <==
---
title: "01-preparacion"
author: "Kevin Carrasco"
date: "2024-01-17"
output:
  rmdformats::readthedown:
    css: color.css
    highlight: tango
    code_folding: hide
    self_contained: true
    thumbnails: false
    lightbox: false
---

```{r setup}
knitr::opts_chunk$set(
	echo = FALSE,
	results = "asis",
	message = FALSE,
	warning = FALSE, 
	fig.height = 18, 
	fig.width = 11, 
	cache = TRUE
)

options(scipen=9999) # desactivar notacion cientifica
remove(list = ls()) #limpieza del entorno de trabajo
```

```{css, echo=FALSE}
# /* expandir el espacio del documento*/
body .main-container {
      max-width: 1600px !important;
      width: 1600px !important;
    }
#/* expandir width de la TOC*/
div.tocify {
    width: 22% !important;
    max-width: 331px !important;
    max-height: 102% !important;
}
```

# packages

```{r packages}
pacman::p_load(dplyr, sjmisc, sjlabelled, summarytools, haven)
```

```{r data}
load("input/data/proc/encuesta_hvs.RData")

#hvs<-readxl::read_xlsx("input/data/proc/encuesta_hvs.xlsx")
#hvs$RBD <- sub("-.*", "", hvs$carac_d2)
#writexl::write_xlsx(hvs, "input/data/proc/encuesta_hvs.xlsx")
```

```{r}
class(hvs)
class(hvs$hva_8)
class(hvs$carac_d6)
class(hvs$RBD)
#sjmisc::frq(hvs$carac_d6)
#sjmisc::frq(hvs$hva_1)
#sjmisc::frq(hvs$hva_8)
#sjmisc::frq(hvs$ha_1)
#sjmisc::frq(hvs$ha_11)
#sjmisc::frq(hvs$eei_1)
#sjmisc::frq(hvs$eei_1_1)
#sjmisc::frq(hvs$ge_1)
```

```{r}
hvs <- hvs %>% dplyr::rename(porcentaje_1= hva_8)
hvs <- hvs %>% dplyr::rename(porcentaje_2= hva_9)
hvs <- hvs %>% dplyr::rename(porcentaje_3= hva_10)

hvs <- hvs %>% dplyr::rename(politic_1= eei_1)
hvs <- hvs %>% dplyr::rename(politic_2= eei_2)

hvs <- hvs %>% 
  mutate_at(vars(starts_with(c("hva_", "ha_", "eei_", "ge_"))), ~ set_na(., na = c(5,6,7,8)))
hvs <- hvs %>% 
  mutate_at(vars(starts_with(c("hva_", "ha_", "eei_", "ge_"))), ~ as_factor(.))

hvs <- hvs %>% 
  mutate_at(vars(starts_with(c("porcentaje_"))), ~ set_na(., na = c(1,2)))
hvs <- hvs %>% 
  mutate_at(vars(starts_with(c("porcentaje_"))), ~ as.numeric(.))

hvs <- hvs %>% 
  mutate_at(vars(starts_with(c("politic_"))), ~ set_na(., na = c(3,4,5,6)))
hvs <- hvs %>% 
  mutate_at(vars(starts_with(c("politic_"))), ~ as_factor(.))

hvs$carac_d4 <- as_factor(hvs$carac_d4)
hvs$carac_d5 <- as.numeric(hvs$carac_d5)
hvs$carac_d6 <- as.numeric(hvs$carac_d6)

sjmisc::frq(hvs$carac_d6)
sjmisc::frq(hvs$hva_1)
sjmisc::frq(hvs$porcentaje_1)
sjmisc::frq(hvs$ha_1)
sjmisc::frq(hvs$ha_11)
sjmisc::frq(hvs$politic_2)
sjmisc::frq(hvs$eei_1_1)
sjmisc::frq(hvs$ge_1)
```

```{r descriptivos}
df1 <- hvs %>% dplyr::select(carac_d4, carac_d5, carac_d6, starts_with("hva"), starts_with("ha"), starts_with("porcentaje"), starts_with("politic"), starts_with("eei"), starts_with("ge")) %>% 
  dfSummary(.,
               plain.ascii = FALSE,
               style = "grid",
               tmp.img.dir = "/tmp",
               graph.magnif = 0.75,
               headings = F,  # encabezado
               varnumbers = F, # num variable
               labels.col = T, # etiquetas
               na.col = T,    # missing
               graph.col = F, # plot
               valid.col = T # n valido
               )
df1$Variable <- NULL
print(df1, method = "render")
```

```{r}
save(hvs, file="input/data/proc/hvs_proc.RData")
```
