==> processing/02-analisis.Rmd <==
---
title: "02-analisis"
author: "Kevin Carrasco"
date: "2024-01-17"
output:
  rmdformats::readthedown:
    css: color.css
    highlight: tango
    code_folding: hide
    self_contained: true
    thumbnails: false
    lightbox: false
---

```{r setup}
knitr::opts_chunk$set(
	echo = FALSE,
	results = "asis",
	message = FALSE,
	warning = FALSE, 
	fig.height = 18, 
	fig.width = 11, 
	cache = TRUE
)

options(scipen=9999) # desactivar notacion cientifica
remove(list = ls()) #limpieza del entorno de trabajo
```

```{css, echo=FALSE}
# /* expandir el espacio del documento*/
body .main-container {
      max-width: 1600px !important;
      width: 1600px !important;
    }
#/* expandir width de la TOC*/
div.tocify {
    width: 22% !important;
    max-width: 331px !important;
    max-height: 102% !important;
}
```

```{r packages}
pacman::p_load(dplyr, ggplot2, sjmisc, sjlabelled, sjPlot, ggrepel, Hmisc, tidyverse, memisc, codebook,
               ggeasy, grid, gridExtra)

```

```{r data, include=FALSE}
load("input/data/proc/hvs_proc.RData")

#hvs<-readxl::read_xlsx("input/data/proc/encuesta_hvs.xlsx")
data <- read_csv2("input/data/original/20220914_Directorio_Oficial_EE_2022_20220430_WEB.csv")
data$RBD <- as.character(data$RBD)
hvs$RBD<-as.factor(hvs$RBD)
test<- left_join(hvs, data, by="RBD")
```
```{r variables, include=FALSE}
names(hvs)
contents(hvs)
get_label(hvs)
get_labels(hvs)
```


```{r include=FALSE}
# Eliminar los asteriscos de las etiquetas

# Obtener nombres de las columnas con etiquetas
columnas_con_etiquetas <- names(hvs)

# Iterar sobre las columnas
for (col in columnas_con_etiquetas) {
  # Verificar si la columna tiene etiquetas
  if (has_label(hvs[[col]])) {
    # Modificar las etiquetas eliminando los asteriscos
    hvs[[col]] <- set_label(hvs[[col]], gsub("\\*", "", label(hvs[[col]])))
  }
}

# Imprimir el resultado
get_label(hvs)


```

# Caracterización

```{r}
test <- test %>% mutate(dependencia = case_when(COD_DEPE2==1 ~ "Municipal",
                                                COD_DEPE2==2 ~ "Part. subvencionado",
                                                COD_DEPE2==3 ~ "Part. pagado",
                                                COD_DEPE2==4 ~ "Adm. delegada",
                                                COD_DEPE2==5 ~ "Serv. local educación",
    is.na(COD_DEPE2) ~ "Municipal"))  # Tratar explícitamente los NA
frq(test$dependencia)
```

```{r}
dependencia = round(prop.table(table(categorias=factor(test$dependencia))),2)
dependencia = as.data.frame(dependencia)

dependencia <- ggplot(dependencia, aes(x = 2, y = -Freq, fill = categorias)) +
  geom_bar(stat = "identity", color = "white") +
  geom_text(aes(label = scales::percent(Freq)), position = position_stack(vjust = 0.5), color = "black", size = 6) +
  coord_polar(theta = "y") +
  scale_fill_brewer(palette = "Set1") +
  theme_void() +  # Cambiar a theme_minimal() para tener un fondo blanco
  theme(legend.title = element_blank()) +
  xlim(0.5, 2.5)

dependencia
ggsave(dependencia, file="output/graphs/dependencia.png", width=8.4, height=8)
```

```{r}
director = round(prop.table(table(categorias=test$carac_d4)),2)
director = as.data.frame(director)
director$categorias <- factor(director$categorias, levels = c(1, 2), labels = c("Director", "Directora"))

director <- ggplot(director,aes(x=2,y=-Freq, fill=categorias))+
  geom_bar(stat = "identity",
           color="white")+
    geom_text(aes(label = scales::percent(Freq)),
              position=position_stack(vjust=0.5),color="black",size=6)+
  coord_polar(theta = "y")+
  scale_fill_brewer(palette = "Set1")+
    theme_void()+
  theme(legend.title = element_blank())+
  xlim(0.5,2.5)
director
ggsave(director, file="output/graphs/director.png", width=8.4, height=8)
```

```{r}
provincia<-test %>% filter(!is.na(NOM_DEPROV_RBD)) %>% 
  group_by(NOM_DEPROV_RBD) %>%
  summarise(cantidad = n()) %>%
  ggplot(aes(x = reorder(NOM_DEPROV_RBD, cantidad), y = cantidad, fill = NOM_DEPROV_RBD)) +
  geom_bar(stat = "identity") +
  geom_text(aes(label = cantidad), vjust = -0.5, color = "black", size = 3) +  # Añade etiquetas de cantidad
  ylab("Cantidad") +
  xlab("Provincia") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_brewer(palette = "Set1") +
  guides(fill = FALSE)

ggsave(provincia, file="output/graphs/provincia.png", width=8.4, height=8)
```

# Hábitos de vida activa

```{r hva1}
plot_hva1 <- sjPlot::plot_stackfrq(dplyr::select(hvs, hva_1, hva_2, hva_3, hva_7),
                      expand.grid =TRUE,
                      geom.colors = 'Spectral',
                      show.total = FALSE, 
                      vjust=rep(c("bottom", "top", "bottom", "top"),4),
#                      sort.frq = "last.desc"
                      ) +
  theme_light() +
  theme(legend.position="bottom", text = element_text(size = 14))

plot_hva1

ggsave(plot_hva1, file="output/graphs/plot_hva1.png", width=8.4, height=8)
```

```{r hva2}
plot_hva2 <- sjPlot::plot_stackfrq(dplyr::select(hvs, hva_4, hva_5, hva_6),
                      expand.grid =TRUE,
                      geom.colors = 'Spectral',
                      show.total = FALSE, 
                      vjust=rep(c("bottom", "top", "bottom", "top"),3),
#                      sort.frq = "last.desc"
                      ) +
  theme_light() +
  theme(legend.position="bottom", text = element_text(size = 14))

plot_hva2

ggsave(plot_hva2, file="output/graphs/plot_hva2.png", width=8.4, height=8)
```


```{r hva porcentajes}
# Promedios de las 3 variables de porcentaje

hvs$promedio1 <-  mean(as.numeric(hvs$porcentaje_1), na.rm = TRUE)
hvs$promedio2 <-  mean(as.numeric(hvs$porcentaje_2), na.rm = TRUE)
hvs$promedio3 <-  mean(as.numeric(hvs$porcentaje_3), na.rm = TRUE)

hvs$promedio1 <- round(hvs$promedio1, digits = 1)
hvs$promedio2 <- round(hvs$promedio2, digits = 1)
hvs$promedio3 <- round(hvs$promedio3, digits = 1)

hvs_long <- tidyr::gather(hvs, key = "variable", value = "porcentaje", promedio1, promedio2, promedio3)

plot_porcentajes <- ggplot(hvs_long, aes(x = variable, y = porcentaje, fill = variable)) +
  geom_bar(stat = "identity", position = "dodge") +
  scale_x_discrete(labels = c("Se trasladan al establecimiento\ncaminando o en bicicleta", "Practican deportes de manera\nextraprogramática en el establecimiento", "Practican deportes de\nmanera extraprogramática\nfuera del establecimiento")) +
  labs(x = 'Hábitos de vida activa de Estudiantes', y = "Porcentaje promedio") + 
  geom_text(aes(label = paste0(porcentaje, "%")), position = "dodge") +
  scale_fill_brewer(palette = "Spectral") +
  theme_light() +
  theme(legend.position="none", text = element_text(size = 14))

plot_porcentajes

ggsave(plot_porcentajes, file="output/graphs/plot_hva_p.png", width=8.4, height=8)
```

# Hábitos alimenticios

```{r ha1_1}
plot_ha1_1 <- sjPlot::plot_stackfrq(dplyr::select(hvs, ha_1, ha_2, ha_3, ha_9),
                      vjust=rep(c("bottom", "top", "bottom", "top"),4),
                      expand.grid =TRUE,
#                      sort.frq = "last.desc",
                      geom.colors = 'Spectral',
                      show.total = FALSE
                      )+
  theme_light() +
  theme(legend.position="bottom", text = element_text(size = 14))

plot_ha1_1

ggsave(plot_ha1_1, file="output/graphs/plot_ha1_1.png", width=8.4, height=8)
```

```{r ha_1_2}
plot_ha1_2 <- sjPlot::plot_stackfrq(dplyr::select(hvs, ha_4, ha_5, ha_6, ha_7, ha_8),
                      vjust=rep(c("bottom", "top", "bottom", "top"),5),
                      expand.grid =TRUE,
#                      sort.frq = "last.desc",
                      geom.colors = 'Spectral',
                      show.total = FALSE
                      )+
  theme_light() +
  theme(legend.position="bottom", text = element_text(size = 14))

plot_ha1_2

ggsave(plot_ha1_2, file="output/graphs/plot_ha1_2.png", width=8.4, height=8)

```


```{r ha2}
frq(hvs$ha_11)
categorias_ha_10 <- c("Nunca", "1 vez a la semana", "2 a 4 veces a la semana", "Todos los días")
# Recodificar ha_11 para que tenga las mismas categorías que ha_10
hvs <- hvs %>%
  mutate(ha_11 = factor(ha_11,
                              levels = c("Nunca", "2 vez a la semana", 
                                         "3 a 4 veces a la semana", "Todos los días"),
                              labels = categorias_ha_10))
hvs$ha_11 <- set_label(x = hvs$ha_11,label = "2.11 Comen comida rápida en su escuela (por ejemplo, sopaipillas, completos, pizzas, hamburguesas, etc.)")


plot_ha2 <- hvs %>% 
  dplyr::select(ha_10, ha_11) %>% 
#  na.omit() %>% 
  sjPlot::plot_stackfrq(., 
                        expand.grid = TRUE, 
                        geom.colors = 'Spectral', 
                        show.total = FALSE) +
  theme_light() +
  theme(legend.position = "bottom", text = element_text(size = 14))

plot_ha2

ggsave(plot_ha2, file="output/graphs/plot_ha2.png", width=8.4, height=8)
```

# Enfoque escolar integral

```{r eei2_part1}
plot_eei2_1 <- sjPlot::plot_stackfrq(dplyr::select(hvs, eei_3, eei_4, eei_5, eei_6),
                      vjust=rep(c("bottom", "top", "bottom", "top"),4),
                      expand.grid =TRUE,
#                      sort.frq = "last.desc",
                      geom.colors = 'Spectral',
                      show.total = FALSE,
                      )+
  theme_light() +
  theme(legend.position="bottom", text = element_text(size = 14))


plot_eei2_1

ggsave(plot_eei2_1, file="output/graphs/plot_eei2_1.png", width=8.5, height=8)
```

```{r eei2_part2}
plot_eei2_2 <- sjPlot::plot_stackfrq(
  dplyr::select(hvs, eei_7, eei_8),
                      vjust=rep(c("bottom", "top", "bottom", "top"),2),
  expand.grid = TRUE,
  geom.colors = 'Spectral',
  show.total = FALSE,
) +
  scale_fill_manual(values = c("#d53e4f", "#fdae61", "#abdda4", "#2b83ba")) +  # Colores rojo, naranja, verde y azul de la paleta "Spectral"
  theme_light() +
  theme(legend.position = "bottom", text = element_text(size = 14))

plot_eei2_2

ggsave(plot_eei2_2, file="output/graphs/plot_eei2_2.png", width=8.4, height=8)
```

```{r eei_2_part3}
plot_eei2_3 <- 
  sjPlot::plot_stackfrq(dplyr::select(hvs, eei_9, eei_10, eei_11, eei_12),
                      vjust=rep(c("bottom", "top", "bottom", "top"),4),
                      expand.grid =TRUE,
#                      sort.frq = "last.desc",
                      geom.colors = 'Spectral',
                      show.total = FALSE,
                      )+
  theme_light() +
  theme(legend.position="bottom", text = element_text(size = 14))


plot_eei2_3

ggsave(plot_eei2_3, file="output/graphs/plot_eei2_3.png", width=8.4, height=8)
```

```{r politic}
plot_politic <- sjPlot::plot_stackfrq(
  dplyr::select(hvs, politic_1, politic_2),
  expand.grid = TRUE,
  geom.colors = 'Spectral',
  show.total = FALSE
) +
  theme_light() +
  theme(legend.position = "bottom", text = element_text(size = 14))

plot_politic

ggsave(plot_politic, file="output/graphs/plot_politic.png", width=8.4, height=8)
```

```{r eei_3_1}
plot_eei_3_1 <- sjPlot::plot_stackfrq(
  dplyr::select(hvs, eei_1_1, eei_2_1),
                      vjust=rep(c("bottom", "top", "bottom", "top"),2),
  expand.grid = TRUE,
  geom.colors = 'Spectral',
  show.total = FALSE
) +
  theme_light() +
  theme(legend.position = "bottom", text = element_text(size = 14))

ggsave(plot_eei_3_1, file = "output/graphs/plot_eei_3_1.png", width = 7, height = 8)
```


# Gestión escolar

```{r ge}
plot_ge <- sjPlot::plot_stackfrq(dplyr::select(hvs, starts_with("ge_")),
                      vjust=rep(c("bottom", "top", "bottom", "top"),4),
#                      sort.frq = "last.desc",
                      expand.grid =TRUE,
                      geom.colors = 'Spectral',
                      show.total = FALSE
                      )+
  theme_light() +
  theme(legend.position="bottom", text = element_text(size = 14))

plot_ge

ggsave(plot_ge, file="output/graphs/plot_ge.png", width=8.4, height=8)
```
